+++ Cargo.toml
@@ -26,13 +26,11 @@ wasmtime-wasi = "0.18.0"
 wasi-common = "0.18.0"
 libc = "0.2.60"
 
-[dev-dependencies]
-test-programs = { path = "crates/test-programs" }
-
 [profile.release]
 codegen-units = 1
 lto = true
 opt-level = 3
 panic = "abort"
 
-[workspace]
+[features]
+integration-tests = []
