--- crates/test-programs/Cargo.toml
@@ -1,17 +1,0 @@
-[package]
-name = "test-programs"
-version = "0.1.0"
-authors = ["Jakub Konka <kubkon@jakubkonka.com>"]
-edition = "2018"
-
-[features]
-test_programs = []
-
-[dev-dependencies]
-ya-runtime-wasi = { path = "../../", version = "0.2.0" }
-anyhow = "1"
-zip = "0.5"
-tempfile = "3.1.0"
-serde_json = "1"
-env_logger = "0.7"
-
