+++ integration-tests/rust-wasi-tutorial/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "ya-wasi-tests"
+name = "rust-wasi-tutorial"
 version = "0.1.0"
 authors = ["Jakub Konka <kubkon@jakubkonka.com>"]
 edition = "2018"
@@ -7,5 +7,3 @@ edition = "2018"
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
-
-[workspace]
