+++ integration-tests/rust-wasi-tutorial/src/main.rs
